+++ dbmodel/utils/fct/gw_fct_setelevfromdem.sql
@@ -134,8 +134,8 @@ BEGIN
 			--loop over selected nodes, intersect node with raster
 			FOR rec IN EXECUTE v_query LOOP
 
-				EXECUTE 'SELECT ST_Value(rast,1,$1,true) FROM ext_raster_dem WHERE id =
-						(SELECT id FROM ext_raster_dem WHERE st_dwithin (envelope, $1, 1) LIMIT 1)'
+				EXECUTE 'SELECT ST_Value(rast,1,$1,true) FROM v_raster_dem WHERE id =
+						(SELECT id FROM v_raster_dem WHERE st_dwithin (envelope, $1, 1) LIMIT 1)'
 					USING rec.the_geom
 					INTO v_elevation;